+++ src/main/gen/dev/kikugie/stonecutter/intellij/lang/psi/StitcherCondition.java
@@ -4,8 +4,10 @@ package dev.kikugie.stonecutter.intellij.lang.psi;
 import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.psi.PsiElement;
+import dev.kikugie.stonecutter.intellij.lang.access.ScopeDefinition;
+import dev.kikugie.stonecutter.intellij.lang.access.ConditionDefinition;
 
-public interface StitcherCondition extends PsiElement {
+public interface StitcherCondition extends ScopeDefinition, ConditionDefinition {
 
   @Nullable
   StitcherExpression getExpression();