+++ src/main/kotlin/dev/kikugie/stonecutter/intellij/lang/access/ConditionDefinition.kt
@@ -1,0 +1,9 @@
+package dev.kikugie.stonecutter.intellij.lang.access
+
+import com.intellij.psi.PsiElement
+import dev.kikugie.stonecutter.intellij.lang.StitcherTokenTypes
+import dev.kikugie.stonecutter.intellij.util.childrenSeqOfType
+
+interface ConditionDefinition : PsiElement {
+    val sugar: Sequence<PsiElement> get() = childrenSeqOfType(StitcherTokenTypes.SUGAR)
+}