+++ src/main/kotlin/dev/kikugie/stonecutter/intellij/service/StonecutterCallbacks.kt
@@ -1,0 +1,78 @@
+package dev.kikugie.stonecutter.intellij.service
+
+import com.intellij.openapi.actionSystem.ActionManager
+import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.editor.Editor
+import com.intellij.openapi.editor.EditorFactory
+import com.intellij.openapi.editor.ex.EditorEventMulticasterEx
+import com.intellij.openapi.editor.ex.FocusChangeListener
+import com.intellij.openapi.fileEditor.FileEditorManager
+import com.intellij.openapi.fileEditor.FileEditorManagerListener
+import com.intellij.openapi.module.ModuleUtilCore
+import com.intellij.openapi.vfs.VirtualFile
+import com.intellij.openapi.vfs.findDocument
+import com.intellij.openapi.vfs.toNioPathOrNull
+import dev.kikugie.stonecutter.intellij.action.VersionSelectorAction
+import dev.kikugie.stonecutter.intellij.editor.StitcherFoldingBuilder.Constants.STITCHER_SCOPE
+import dev.kikugie.stonecutter.intellij.settings.StonecutterSettings
+import java.awt.event.FocusEvent
+
+object StonecutterCallbacks {
+    internal fun invokeAppLoad(settings: StonecutterSettings) {
+        (EditorFactory.getInstance().eventMulticaster as? EditorEventMulticasterEx)
+            ?.addFocusChangeListener(createFocusListener(), settings)
+    }
+
+    internal fun invokeProjectLoad(service: StonecutterService) {
+        val bus = service.project.messageBus.connect(service)
+        bus.subscribe(FileEditorManagerListener.FILE_EDITOR_MANAGER, createFileListener(service))
+    }
+
+    internal fun invokeProjectReload(service: StonecutterService) {
+        (ActionManager.getInstance().getAction("dev.kikugie.stonecutter.intellij.select_version") as VersionSelectorAction)
+            .isAvailable = true
+    }
+
+    //<editor-fold desc="Implementations">
+    /**
+     * Update folding for blocks commented out by Stonecutter.
+     * Can be disabled in code folding settings.
+     */
+    private fun createFocusListener() = object : FocusChangeListener {
+        override fun focusGained(editor: Editor, event: FocusEvent) {
+            if (!StonecutterSettings.foldDisabledScopes) return
+            if (!event.isTemporary) editor.foldingModel.runBatchFoldingOperation {
+                for (it in editor.foldingModel.allFoldRegions) if (it.group == STITCHER_SCOPE)
+                    it.isExpanded = false
+            }
+        }
+    }
+
+    /**
+     * Disallow editing generated Stonecutter files.
+     */
+    private fun createFileListener(service: StonecutterService) = object : FileEditorManagerListener {
+        override fun fileOpened(source: FileEditorManager, file: VirtualFile) {
+            val application = ApplicationManager.getApplication()
+            application.invokeLater {
+                application.runWriteAction {
+                    updateFile(service, source, file)
+                }
+            }
+        }
+
+        // TODO: Make a setting for it
+        private fun updateFile(service: StonecutterService, source: FileEditorManager, file: VirtualFile) {
+            val path = file.toNioPathOrNull() ?: return
+            val module = ModuleUtilCore.findModuleForFile(file, source.project)
+            if (module != null) return
+            val root = service.lookup.all
+                .find { path.startsWith(it.location.resolve("build")) }
+                ?: return
+            val isGenerated = path.startsWith(root.location.resolve("build/generated/stonecutter"))
+            val isCache = path.startsWith(root.location.resolve("build/stonecutter-cache"))
+            if (isGenerated || isCache) file.findDocument()?.setReadOnly(true)
+        }
+    }
+    //</editor-fold>
+}